+++ src/bus_i2c.cpp
@@ -55,7 +55,7 @@ BusI2CDevice::BusI2CDevice(TwoWire &bus, uint32_t address) : bus_(bus)
     address_ = address;
 }
 
-size_t BusI2CDevice::writable()
+int32_t BusI2CDevice::writable()
 {
     return 32; // Maybe change this to something else. arduino might support availableForWrite() or something.
 }
